section 6,7 end
