+++ src/main/java/com/example/demo/app/inquiry/InquiryController.java
@@ -2,18 +2,19 @@ package com.example.demo.app.inquiry;
 
 import java.time.LocalDateTime;
 import java.util.List;
-
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
 import com.example.demo.entity.Inquiry;
+import com.example.demo.service.InquiryNotFoundException;
 import com.example.demo.service.InquiryService;
 
 @Controller
@@ -22,6 +23,7 @@ public class InquiryController {
 
 	private final InquiryService inquiryService;
 
+	@Autowired
 	public InquiryController(InquiryService inquiryService){
 		this.inquiryService = inquiryService;
 	}
@@ -30,10 +32,30 @@ public class InquiryController {
 	public String index(Model model) {
 		List<Inquiry> list = inquiryService.getAll();
 
-		model.addAttribute("inquiryList", list);
-		model.addAttribute("title", "Inquiry Index");
+//		--------2021/12/28 section7 Delete Start↓---------
+//		model.addAttribute("inquiryList", list);
+//		model.addAttribute("title", "Inquiry Index");
+//		
+//		return "inquiry/index";
+//      --------2021/12/28 section7 Delete End↑---------
+		
+//		return "inquiry/index_boot";
+//		
+//		Inquiry inquiry = new Inquiry();
+//		inquiry.setId(4);
+//		inquiry.setName("Kazuhiro");
+//		inquiry.setEmail("kazuhiro-kawase@example.com");
+//		inquiry.setContents("こんにちは");
+		
+//		try {
+//		  inquiryService.update(inquiry);
+//		} catch (InquiryNotFoundException e) {
+//		  model.addAttribute("message", e);
+//		  return "error/CustomPage";
+//		}
 
 		return "inquiry/index_boot";
+		
 	}
 
 	@GetMapping("/form")
@@ -85,12 +107,15 @@ public class InquiryController {
 		//演習1-2-6 completeメソッド内で、inquiryForm.getId()が0の場合はinquiryServiceのsaveメソッドを実行
 		//0ではない場合はidをセットしinquiryServiceのupdateメソッドをかける
 		if(inquiryForm.getId() == 0) {
-			inquiryService.save(inquiry);
+//			inquiryService.save(inquiry);
 		} else {
-			inquiry.setId(inquiryForm.getId());
-			inquiryService.update(inquiry);
+//			section  6 update
+//          inquiry.setId(inquiryForm.getId());
+//			inquiryService.update(inquiry);
 		}
 
+//      section  6 update
+		inquiryService.save(inquiry);
 		redirectAttributes.addFlashAttribute("complete", "Completed!");
 		return "redirect:/inquiry/form";
 	}
@@ -113,4 +138,11 @@ public class InquiryController {
 		//演習1-2-5 フォーム画面へ遷移させる
 		return "inquiry/form_boot";
 	}
+	
+	//section7 1つのコントローラ内で共通の例外処理
+//	@ExceptionHandler(InquiryNotFoundException.class)
+//	public String handleException(InquiryNotFoundException e, Model model) {
+//	  model.addAttribute("massage", e);
+//	  return "error/CustomPage";
+//	}
 }
