+++ src/main/java/com/example/demo/dao/InquiryDao.java
@@ -1,0 +1,13 @@
+package com.example.demo.dao;
+
+import java.util.List;
+import com.example.demo.entity.Inquiry;
+
+public interface InquiryDao {
+  void insertInquiry(Inquiry inquiry);
+  
+  int updateInquiry(Inquiry inquiry);
+  
+  List<Inquiry> getAll();
+
+}
