+++ src/main/java/com/example/demo/repository/InquiryDao_bk.java
@@ -1,0 +1,13 @@
+//package com.example.demo.repository;
+//
+//import java.util.List;
+//import com.example.demo.entity.Inquiry;
+//
+//public interface InquiryDao_bk {
+//	
+//	void insertInquiry(Inquiry inquiry);
+//	
+//	int updateInquiry(Inquiry inquiry);
+//	
+//	List<Inquiry> getAll();
+//
